[src] Float sample rate

+++ asteroid/filterbanks/analytic_free_fb.py
@@ -15,7 +15,7 @@ class AnalyticFreeFB(Filterbank):
         kernel_size (int): Length of the filters.
         stride (int, optional): Stride of the convolution.
             If None (default), set to ``kernel_size // 2``.
-        sample_rate (int): Sample rate of the expected audio.
+        sample_rate (float): Sample rate of the expected audio.
             Defaults to 8000.
 
     Attributes:
@@ -27,7 +27,7 @@ class AnalyticFreeFB(Filterbank):
         Antoine Deleforge, Emmanuel Vincent.
     """
 
-    def __init__(self, n_filters, kernel_size, stride=None, sample_rate=8000, **kwargs):
+    def __init__(self, n_filters, kernel_size, stride=None, sample_rate=8000.0, **kwargs):
         super().__init__(n_filters, kernel_size, stride=stride, sample_rate=sample_rate)
         self.cutoff = int(n_filters // 2)
         self.n_feats_out = 2 * self.cutoff
