+++ asteroid/filterbanks/enc_dec.py
@@ -13,14 +13,14 @@ class Filterbank(nn.Module):
         kernel_size (int): Length of the filters.
         stride (int, optional): Stride of the conv or transposed conv. (Hop size).
             If None (default), set to ``kernel_size // 2``.
-        sample_rate (int): Sample rate of the expected audio.
+        sample_rate (float): Sample rate of the expected audio.
             Defaults to 8000.
 
     Attributes:
         n_feats_out (int): Number of output filters.
     """
 
-    def __init__(self, n_filters, kernel_size, stride=None, sample_rate=8000):
+    def __init__(self, n_filters, kernel_size, stride=None, sample_rate=8000.0):
         super(Filterbank, self).__init__()
         self.n_filters = n_filters
         self.kernel_size = kernel_size
